refactor(runtime-vapor): create component slots
