+++ packages/runtime-vapor/__tests__/componentAttrs.spec.ts
@@ -41,6 +41,7 @@ describe('attribute fallthrough', () => {
             },
           ],
           null,
+          null,
           true,
         )
       },
@@ -87,6 +88,7 @@ describe('attribute fallthrough', () => {
             },
           ],
           null,
+          null,
           true,
         )
       },
@@ -126,6 +128,7 @@ describe('attribute fallthrough', () => {
             },
           ],
           null,
+          null,
           true,
         )
         return n0
@@ -147,6 +150,7 @@ describe('attribute fallthrough', () => {
               id: () => _ctx.id,
             },
           ],
+          null,
           null,
           true,
         )
