+++ packages/runtime-vapor/src/apiCreateComponent.ts
@@ -5,13 +5,14 @@ import {
 } from './component'
 import { setupComponent } from './apiRender'
 import type { RawProps } from './componentProps'
-import type { Slots } from './componentSlots'
+import type { DinamicSlotsGetter, Slots } from './componentSlots'
 import { withAttrs } from './componentAttrs'
 
 export function createComponent(
   comp: Component,
   rawProps: RawProps | null = null,
   slots: Slots | null = null,
+  dynamicSlotsGetter: DinamicSlotsGetter | null = null,
   singleRoot: boolean = false,
 ) {
   const current = currentInstance!
@@ -19,6 +20,7 @@ export function createComponent(
     comp,
     singleRoot ? withAttrs(rawProps) : rawProps,
     slots,
+    dynamicSlotsGetter,
   )
   setupComponent(instance, singleRoot)
